+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## [Unreleased]
+### Changed
+- entropy(): renamed argument 'dist' to 'plugin'
 
 ## [0.6] - 2018-03-22
 ### Changed